Update README and enhance calendar functionality
A promo image has been added to README and the functionality of the JQUery and Bootstrap 5 calendar has been enhanced. Updates include changes to multiple method names from FontAwesome to Bootstrap Icons and the addition of a dark mode switch. Calendar now also includes an ability to change language settings.