Add popover feature displaying daily appointments
A new popover feature has been added that pops up to display daily appointments. A popover configuration attribute has been included that allows the customization of the popover. This new functionality is only active when the "showPopover" attribute is set to true. The jQuery bootstrap calendar has been updated to accommodate this new feature.