Implement forced date selection in JS calendar
This change introduces a new functionality that enables forced date selection in the JavaScript calendar. A boolean variable, forceDate, is added to handle the presence or absence of a selected date. If no date is selected, the calendar will automatically select the current system date. If a date is selected, it is formatted and triggered to be selected in the calendar. This commit enhances user experience by providing an automatic selection mechanism when no specific date is selected by the user.