Add new jQuery bootstrap calendar
The bsCalendar jQuery plugin was implemented. This plugin is capable of creating a calendar and populating it with events. The plugin supports a variety of options and callback functions providing full control over the appearance and behavior of the calendar. It makes AJAX requests to load data from a server and can format and display a wide range of events. Additionally, the plugin also extends the native JavaScript Date object to offer a more robust tool for manipulating dates.